pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "diaricat-live"
version = "0.1.0"
description = "Live audio stream transcription and alert service for Nyx Terminal"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Nia Huck" }]

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "yt-dlp>=2024.0",
    "sse-starlette>=2.0",
    "faster-whisper>=1.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest",
    "pytest-asyncio",
    "httpx",
]

[project.scripts]
diaricat-live = "diaricat_live.run:main"

[tool.hatch.build.targets.wheel]
packages = ["src/diaricat_live"]

[tool.ruff]
target-version = "py311"
line-length = 100
